2 course/4 term/PTI/LR-2/requirements.txt
setuptools~=59.4.0
toml~=0.10.2
PyYAML~=6.0